chat/requirements.txt
transformers
tokenizers
deepspeed==0.9.1
datasets
accelerate
pytorch
wandb